requirements/base.txt
Django==1.11.5
dj-database-url==0.4.1
django-moj-irat==0.3
django-rest-swagger==2.1.2
djangorestframework>=3.7.0,<3.8.0
django-extended-choices==1.1.2
django-cors-headers==2.1.0
django-filter==1.0.4
drf-nested-routers==0.90.0
flake8==3.4.1
openpyxl==2.4.8
psycopg2==2.6.2
PyYAML==3.12
six==1.11.0
uWSGI==2.0.14
xlrd==1.0.0
